Fixed masthead
Masthead images are blocking the Tezbox logo as well as the navigation menu on screens with 992px or more (my laptop) and it's also omitting the words "with macos, Windows, *nix and iOS in development..." from under the Google Play icon. 

After the fix the background images are in the proper place and the text below the Google Play icon is back in view.